# Use `command -v` to reduce zsh startup time