`which` is very slow.

Simple benchmark:
```
# before
$ time ( for i in {1..100}; do zsh z.lua.plugin.zsh; done )
18.54s user 2.29s system 102% cpu 20.340 total
# after
$ time ( for i in {1..100}; do zsh z.lua.plugin.zsh; done )
0.84s user 0.37s system 104% cpu 1.164 total
```

22 times faster!